refactor(desktop): extract App support functions into app_support.rs (WIP)
Move the standalone helpers below the App component (gateway connection +
event application, DOM query, directory helpers, swarm ops, the event buffer)
out of app.rs into app_support.rs. app.rs drops from 1992 to 1418 lines.

Still over 1000: the App component's ~788-line rsx! render tree and its event
handlers remain to be decomposed in a follow-up.
